+++ sdk/core/test/unit/CommonNetworkMember.test.ts
@@ -42,6 +42,7 @@ import cognitoUserTokens from '../factory/cognitoUserTokens'
 import credentialShareRequestToken from '../factory/credentialShareRequestToken'
 import parsedCredentialShareRequestToken from '../factory/parsedCredentialShareRequestToken'
 import parsedCredentialShareResponseToken from '../factory/parsedCredentialShareResponseToken'
+import presentationWithSubmission from '../factory/presentationWithSubmission'
 import { testPlatformTools } from '../helpers/testPlatformTools'
 import { randomBytes } from '../../src/shared/randomBytes'
 
@@ -1752,6 +1753,32 @@ describe('CommonNetworkMember', () => {
     } else {
       expect(result.suppliedPresentation).to.deep.eq(vp)
       expect(result.errors.join('\n')).to.contains('The challenge is not as expected')
+    }
+  })
+
+  it('#verifyPresentation should work with presentation that has submissions', async () => {
+    const requesterCommonNetworkMember = new AffinidiWallet(walletPassword, encryptedSeedElem, options)
+    const userSeed =
+      'a6a6c462b15619273fbf942c99e2b7c1be03ef184f5d1d27a17f29b45c4676dcc5e07ecc835acf1ff1d6f437e95f5cc77ce597d254a7b7' +
+      'db932bbb2698d0804e47cd02c7f02580da2132d008c0e313820964646c7630e01d65925d4b5f8db741'
+    const userPassword = 'Test8b60806087f5e7b8eee6607ae8!!'
+    const userCommonNetworkMember = new AffinidiWallet(userPassword, userSeed, options)
+
+    const presentationChallenge = 'beb140f5-e746-4ba0-8bd2-f6ecf26c3f25'
+    const domain = 'domain'
+    const vp = await userCommonNetworkMember.signUnsignedPresentation(
+      presentationWithSubmission,
+      presentationChallenge,
+      domain,
+    )
+    const result = await requesterCommonNetworkMember.verifyPresentation(vp)
+
+    if (result.isValid === true) {
+      expect(result.challenge).to.eq(presentationChallenge)
+      expect(result.suppliedPresentation).to.deep.eq(vp)
+    } else {
+      expect(result.suppliedPresentation).to.deep.eq(vp)
+      expect.fail(result.errors.join('\n'))
     }
   })