+++ sdk/node/CHANGELOG.md
@@ -1,4 +1,5 @@
-# release 7.13.1 (2023-03-09)
+# release 7.13.2 (2023-03-10)
+* Update common libs to support presentation submission v2 # release 7.13.1 (2023-03-09)
 * updated `validateJWT` error handling
 # release 7.13.0 (2023-03-08)
 * updated to use latest common lib with KeyManager