+++ sdk/vc-data/CHANGELOG.md
@@ -1,3 +1,5 @@
+## release 1.6.2 (2023-03-10)
+* bumped `vc-common` lib version.
 ## release 1.6.1 (2023-02-09)
 * bumped `vc-common` lib version.